Fix module structure and resolve compilation errors
- Add module declarations to `src/lib.rs` for `types`, `core`, `detectors`, and `math`.
- Remove non-existent and unused imports in `src/lib.rs` and `src/meta_cognitive.rs`.
- Fix ownership and type mismatch issues in `src/lib.rs` and `src/main.rs`.
- Ensure proper trait scoping for `chrono::Timelike` in `src/detectors.rs`.
- Add `.gitignore` to exclude build artifacts.
- Regenerate `Cargo.lock` correctly via `cargo build`.

+++ src/main.rs
@@ -52,7 +52,7 @@ fn generate_mock_data(count: usize) -> Vec<Candle> {
         price += noise;
         
         data.push(Candle {
-            timestamp: now + Duration::hours(i),
+            timestamp: now + Duration::hours(i as i64),
             open: price,
             high: price + 0.0005,
             low: price - 0.0005,
